% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/helpers.R
\name{hcrParams}
\alias{hcrParams}
\alias{hcrParams,FLQuants-method}
\alias{hcrParams,list-method}
\title{Build ICES hockey-stick HCR parameters}
\usage{
hcrParams(object, ...)

\S4method{hcrParams}{FLQuants}(object, fmin = 0.005, bmin = 0, ...)

\S4method{hcrParams}{list}(object, fmin = 0.005, bmin = 0, ...)
}
\arguments{
\item{object}{An \code{FLQuants} or \code{list} of reference points.}

\item{...}{Passed to methods.}

\item{fmin}{Multiplier or value for \code{fmin} relative to \code{ftar}
when built from named reference \code{FLQuants} (default 0.005).}

\item{bmin}{Value for \code{bmin} (default 0).}
}
\value{
An \code{FLPar} with params \code{ftar}, \code{btrig}, \code{fmin},
\code{bmin}, \code{blim}.
}
\description{
Generic for constructing \code{FLPar} HCR parameters
(\code{ftar}, \code{btrig}, \code{fmin}, \code{bmin}, \code{blim}) from
reference-point objects.
}
\details{
Expected names in \code{FLQuants} / list: \code{fmsy} (or \code{ftar}),
\code{msybtrigger} (or \code{btrig}), \code{blim}.
}
