% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/helpers.R
\name{lorenzenM}
\alias{lorenzenM}
\alias{lorenzenM,FLQuant-method}
\alias{lorenzenM,FLStock-method}
\title{Lorenzén mass-at-age natural mortality}
\usage{
lorenzenM(object, ...)

\S4method{lorenzenM}{FLQuant}(object, params = FLPar(m1 = 0.15/(0.3^-0.288), m2 = -0.288), ...)

\S4method{lorenzenM}{FLStock}(
  object,
  params = FLPar(m1 = 0.15/(0.3^-0.288), m2 = -0.288),
  replaceAge0 = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{FLQuant} of weights, or an \code{FLStock}.}

\item{...}{Passed to methods (\code{params} as \code{FLPar} with \code{m1},
\code{m2}).}

\item{params}{\code{FLPar} with \code{m1} and \code{m2}.}

\item{replaceAge0}{If \code{TRUE}, set age-0 stock weights from catch weights
before computing M.}
}
\value{
An \code{FLQuant} of natural mortality.
}
\description{
Generic for \eqn{M = m_1 W^{m_2}} natural mortality from mass-at-age.
}
