% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/helpers.R
\name{recDevs}
\alias{recDevs}
\alias{recDevs,FLQuant-method}
\alias{recDevs,FLSR-method}
\title{Simulate recruitment deviations}
\usage{
recDevs(object, ...)

\S4method{recDevs}{FLQuant}(object, nits = 100, trend = FALSE, years = NULL, ...)

\S4method{recDevs}{FLSR}(object, nits = 100, trend = FALSE, years = NULL, ...)
}
\arguments{
\item{object}{An \code{FLQuant} of log residuals, or an \code{FLSR}.}

\item{...}{Passed to methods.}

\item{nits}{Number of iterations.}

\item{trend}{If \code{TRUE}, add a lowess trend (epistemic); if
\code{FALSE}, return AR(1) residuals about the trend (aleatory).}

\item{years}{Optional year window for the ARIMA / AR(1) fit.}
}
\value{
An \code{FLQuant} of log recruitment deviations. Use
\code{exp(recDevs(...))} as \code{sr_deviances} in \code{hcrICES}.
}
\description{
Generic for simulating aleatory / epistemic recruitment deviations from
stock–recruit residuals.
}
