% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/bias.R
\name{retroBias}
\alias{retroBias}
\alias{retroBias,list-method}
\alias{retroBias,FLStocks-method}
\alias{biasFromRetros}
\title{Retrospective assessment bias}
\usage{
retroBias(object, ...)

\S4method{retroBias}{list}(object, years = 2006:2020, ...)

\S4method{retroBias}{FLStocks}(object, years = 2006:2020, ...)

biasFromRetros(retros, years = 2006:2020, ...)
}
\arguments{
\item{object}{A list of peels (\code{FLStocks}) or a list of such lists by OM.}

\item{...}{Passed to methods.}

\item{years}{Integer vector of peel terminal years (default \code{2006:2020}).}
}
\value{
A \code{data.frame} with columns \code{.id}, \code{iYr}, \code{now},
\code{then}.
}
\description{
Generic for computing retrospective SSB bias from peels. Named
\code{retroBias} to avoid masking \code{FLCore::bias}.
}
