% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/samIndices.R
\name{samIndexRec}
\alias{samIndexRec}
\alias{samIndexRec,FLStock-method}
\title{SAM recruitment index from an FLStock}
\usage{
samIndexRec(object, ...)

\S4method{samIndexRec}{FLStock}(
  object,
  name = "Recs",
  desc = "Index equal to Recruitments",
  startf = 0,
  endf = 0.01,
  age = 1,
  ...
)
}
\arguments{
\item{object}{An \code{FLStock}.}

\item{...}{Passed to methods.}

\item{name}{Index name.}

\item{desc}{Index description.}

\item{startf, endf}{Survey timing fractions.}

\item{age}{Age label for recruitment.}
}
\description{
SAM recruitment index from an FLStock
}
