% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/samIndices.R
\name{samIndexSSB}
\alias{samIndexSSB}
\alias{samIndexSSB,FLStock-method}
\title{SAM SSB index from an FLStock}
\usage{
samIndexSSB(object, ...)

\S4method{samIndexSSB}{FLStock}(
  object,
  name = "SSB",
  desc = "Index equal to SSB",
  startf = 0,
  endf = 0.01,
  age = 1,
  ...
)
}
\arguments{
\item{object}{An \code{FLStock}.}

\item{...}{Passed to methods.}

\item{name}{Index name.}

\item{desc}{Index description.}

\item{startf, endf}{Survey timing fractions.}

\item{age}{Age label for the aggregated index.}
}
\description{
SAM SSB index from an FLStock
}
