% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samIndices.R
\name{validateSamIndex}
\alias{validateSamIndex}
\title{Validate FLIndex structure for SAM}
\usage{
validateSamIndex(idx)
}
\arguments{
\item{idx}{An \code{FLIndex} or \code{FLIndices}.}
}
\value{
\code{TRUE} if valid.
}
\description{
Validate FLIndex structure for SAM
}
